test: split metadata test into serialisation cases
